requirements.txt
aiogram==3.0.0
sqlalchemy
psycopg2-binary
python-dotenv
alembic 